requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
python-tds==1.12.0
pydantic==2.9.2